# desktop/src/openapi/spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Aether",
    "description": "something something app",
    "contact": {
      "name": "you"
    },
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/trash/tasks": {
      "get": {
        "tags": [
          "Trash"
        ],
        "summary": "Get all trashed tasks",
        "operationId": "get_trashed_tasks",
        "responses": {
          "200": {
            "description": "List of deleted tasks",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Task"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/v1/trash/{id}/restore": {
      "post": {
        "tags": [
          "Trash"
        ],
        "summary": "Restore a deleted task",
        "operationId": "restore_task",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Task ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Task restored"
          },
          "404": {
            "description": "Task not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Task": {
        "type": "object",
        "description": "Task model",
        "required": [
          "id",
          "title",
          "isCompleted",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "dueDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "goalId": {
            "type": "string",
            "nullable": true
          },
          "goalInstanceId": {
            "type": "string",
            "nullable": true
          },
          "id": {
            "type": "string"
          },
          "isCompleted": {
            "type": "boolean"
          },
          "title": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
...
}